added role based access control added logging finished access management theres still a bug on the admin page where it always says there are no lottery winners even if there should be

+++ lottery/views.py
@@ -2,24 +2,26 @@
 import copy
 import logging
 from flask import Blueprint, render_template, request, flash
-from app import db
+from app import db, requires_roles
 from models import Draw, User
 from flask_login import login_required, current_user
 
 # CONFIG
 lottery_blueprint = Blueprint('lottery', __name__, template_folder='templates')
 
+
 # VIEWS
 # view lottery page
 @lottery_blueprint.route('/lottery')
 @login_required
+@requires_roles('user')
 def lottery():
-
     return render_template('lottery.html')
 
 
 @lottery_blueprint.route('/add_draw', methods=['POST'])
 @login_required
+@requires_roles('user')
 def add_draw():
     submitted_draw = ''
     for i in range(6):
@@ -41,6 +43,7 @@ def add_draw():
 # view all draws that have not been played
 @lottery_blueprint.route('/view_draws', methods=['POST'])
 @login_required
+@requires_roles('user')
 def view_draws():
     # get all draws that have not been played [played=0]
 
@@ -57,6 +60,7 @@ def view_draws():
 # view lottery results
 @lottery_blueprint.route('/check_draws', methods=['POST'])
 @login_required
+@requires_roles('user')
 def check_draws():
     # get played draws
     played_draws = decrypt_draws(True)
@@ -74,8 +78,9 @@ def check_draws():
 # delete all played draws
 @lottery_blueprint.route('/play_again', methods=['POST'])
 @login_required
+@requires_roles('user')
 def play_again():
-    delete_played = Draw.__table__.delete().where(Draw.played)  # TODO: delete played draws for current user only
+    delete_played = Draw.__table__.delete().where(Draw.played).where(Draw.user_id==current_user.id)
     db.session.execute(delete_played)
     db.session.commit()
